+++ src/components/agents/AgentTerminal.vue
@@ -621,7 +621,9 @@ export default {
       config = { print: true, attempts: 30, delay: 5000 },
     ) {
       if (!config.attempts) config.attempts = 30;
-      if (!config.delay) config.delay = this.agent.delay != null ? this.agent.delay * 1000 : 5000;
+      if (!config.delay)
+        config.delay =
+          this.agent.delay != null ? this.agent.delay * 1000 : 5000;
 
       let res = null;
       let hasPrintedJobStarted = false;