+++ src/components/plugins/PluginTasksTable.vue
@@ -62,9 +62,11 @@
               "
             >
               {{
-                addBlankLines(item.expandedInput
-                  ?
-                  expandedTasks[item.uniqueId].full_input : item.input)
+                addBlankLines(
+                  item.expandedInput
+                    ? expandedTasks[item.uniqueId].full_input
+                    : item.input,
+                )
               }}
             </p>
             <p><b>Task Output:</b></p>
@@ -108,9 +110,7 @@
                 v-if="expandedTasks[item.uniqueId].htmlOutput"
                 v-html="expandedTasks[item.uniqueId].htmlOutput"
               />
-              <div v-else>
-                addBlankLines(item.output)
-              </div>
+              <div v-else>addBlankLines(item.output)</div>
             </div>
           </div>
         </td>
@@ -350,7 +350,7 @@ export default {
       return new AnsiUp().ansi_to_html(output);
     },
     addBlankLines(text) {
-      return `\n${  text  }\n`;
+      return `\n${text}\n`;
     },
     deleteTag(task, tag) {
       pluginApi