+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Removed extra space being added in front of Task Input and Output by append/prepend new line
+-   Changed default sort on agents page to be "first seen" instead of "last seen"
 
 ## [2.7.2] - 2024-01-31